Added a skybox and changed so you can't help multiple blobs at a time

+++ Blobtricity/Assets/Dennis/Scripts/LevelManager.cs
@@ -20,7 +20,7 @@ public class LevelManager : MonoBehaviour
     [Header("Other")]
     [HideInInspector] public UnityStandardAssets.Characters.FirstPerson.MouseLook mouseManager;
 
-    public void FinishLevel(int _energyReduced)
+    public void FinishLevel(float _energyReduced)
     {
         uiCamera.SetActive(true);
         playerGameObject.SetActive(false);
